LayoutTests/imported/w3c/web-platform-tests/websockets/Secure-Send-65K-data.any.worker-expected.txt


FAIL Send  65K data on a Secure WebSocket - Connection should be opened assert_equals: expected 0 but got 65000
PASS Send 65K data on a Secure WebSocket - Message should be received
FAIL Send 65K data on a Secure WebSocket - Connection should be closed assert_true: WebSocket connection should be open expected true got false
